More consistent answers with the requirements + better migration

+++ project/main.py
@@ -5,7 +5,7 @@ from sqlalchemy import text
 from models import *
 from database import get_session, init_models, engine, SessionDep
 from authorization import oauth2_scheme, get_current_user, query_user, verify_password, get_password_hash, create_access_token
-
+from fastapi.responses import JSONResponse
 app = FastAPI()
 
 
@@ -18,12 +18,28 @@ async def on_startup():
 async def on_shutdown():
     await engine.dispose()
 
+forbidden = JSONResponse(status_code=403, content={"message": "Forbidden for you"})
 
-# Module 1
+def generate_validation_error_for_fields(*fields: str):
+    details = {"message": "Validation error",}
+
+    for field in fields:
+        details[field] = "Validation error"
+
+    return JSONResponse(status_code=422, content=details)
+
+def validate_email(email: str | None):
+    if email is None or re.match(r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$', email) is None:
+        return False
+
+    return True
 
 @app.get("/api/health")
 def health_check():
     return {"status": "healthy"}
+
+
+# Module 1
 
 
 @app.post("/login")
@@ -31,13 +47,15 @@ async def login(new_user: UserAuthorization, session: SessionDep):
     user = await query_user(new_user.email, session)
 
     if not user or not verify_password(new_user.password, user.password_hash):
-        raise HTTPException(status_code=401, detail="Invalid credentials")
+        raise HTTPException(status_code=401, detail={
+            "password": "Login failed"
+        })
 
     access_token = create_access_token(data={"sub": user.id})
 
     await session.execute(text("insert into sessions (user_id, token) values (:user_id, :token) returning token"), {"user_id": user.id, "token": access_token})
     await session.commit()
-    return {"message": "Logged in successfully", "access_token": access_token}
+    return {"message": "Logged in successfully", "user_token": access_token}
 
 
 @app.post("/logout")
@@ -50,15 +68,15 @@ async def logout(token: Annotated[str, Depends(oauth2_scheme)], session: Session
 @app.post("/signup")
 async def register(new_user: UserSignup, session: SessionDep):
     try:
-        if re.match(r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$', new_user.email) is None:
-            raise HTTPException(status_code=422, detail={"message": "Validation error", "email": "Validation error", })
+        if not validate_email(new_user.email):
+            return generate_validation_error_for_fields("email")
 
-        if new_user.password is None:
-            raise HTTPException(status_code=422, detail={"message": "Validation error", "password": "Validation error", })
+        if new_user.password == "":
+            return generate_validation_error_for_fields("password")
 
         fio = new_user.fio.split()
         if len(fio) < 2 or len(fio) > 3:
-            raise HTTPException(status_code=422, detail={"message": "Validation error", "fio": "Validation error", })
+            return generate_validation_error_for_fields("fio")
 
         hashed_password = get_password_hash(new_user.password)
 
@@ -74,25 +92,24 @@ async def register(new_user: UserSignup, session: SessionDep):
         id = await session.execute(insert_user, {"name": name, "surname": surname, "middle_name": middle_name, "email": new_user.email, "password": hashed_password})
         await session.commit()
 
+        user_id = id.scalar_one()
+
+        access_token = create_access_token(data={"sub": user_id})
+
+        await session.execute(text("insert into sessions (user_id, token) values (:user_id, :token) returning token"), {"user_id": user_id, "token": access_token})
+        await session.commit()
+
     except Exception as e:
         await session.rollback()
         raise HTTPException(status_code=500, detail=str(e))
 
-    user_id = id.scalar_one()
 
-    access_token = create_access_token(data={"sub": user_id})
-
-    await session.execute(text("insert into sessions (user_id, token) values (:user_id, :token) returning token"), {"user_id": user_id, "token": access_token})
-    await session.commit()
-
-    return {"message": "Registered successfully", "token": access_token}
+    return {"message": "Registered successfully", "user_token": access_token}
 
 
 @app.get("/profile")
 def get_profile(user: Annotated[User, Depends(get_current_user)], session: SessionDep):
-    if not user:
-        raise HTTPException(status_code=404, detail="Not found")
-    return user._asdict()
+    return {"user": {"id": user.id, "fio": user.name + " " + user.surname + " " + (user.middle_name if user.middle_name is not None else ""), "avatar": user.avatar, "email": user.email}}
 
 
 @app.get("/products")
@@ -133,7 +150,7 @@ async def remove_from_cart(cart_id: int, user: Annotated[User, Depends(get_curre
 @app.get("/cart")
 async def get_cart(user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     try:
-        cart_items = (await session.execute(text("select * from cart where user_id = :user_id"), {"user_id": user.id})).all()
+        cart_items = (await session.execute(text("select cart.id, cart.product_id, products.name, products.description, products.price from cart inner join products on cart.product_id = products.id where cart.user_id = :user_id"), {"user_id": user.id})).all()
         return [item._asdict() for item in cart_items]
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
@@ -142,13 +159,16 @@ async def get_cart(user: Annotated[User, Depends(get_current_user)], session: Se
 @app.post("/order")
 async def place_order(user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     try:
-        order_id = await session.execute(text("insert into orders (user_id) values (:user_id) returning id"), {"user_id": user.id})
+        cart_items_quantity = (await session.execute(text("select count(product_id) as quantity from cart where user_id = :user_id"), {"user_id": user.id})).scalar()
+
+        if cart_items_quantity <= 0:
+            raise HTTPException(status_code=400, detail={"message": "Cart is empty"})
+
+        order_id = await session.execute(text("insert into orders (user_id, status, order_price) values (:user_id, 'pending', (select sum(price) from cart inner join products on cart.product_id = products.id where cart.user_id = :user_id)) returning id"), {"user_id": user.id})
+        order_id = order_id.scalar_one()
         await session.commit()
 
-        order_id = order_id.scalar_one()
-
-        await session.execute(text("insert into order_items (order_id, product_id, quantity) select :order_id, product_id, quantity from cart where user_id = :user_id"),
-                              {"order_id": order_id, "user_id": user.id})
+        await session.execute(text("insert into order_items (order_id, product_id) select :order_id, product_id from cart where user_id = :user_id"),{"order_id": order_id, "user_id": user.id})
         await session.execute(text("delete from cart where user_id = :user_id"), {"user_id": user.id})
 
         await session.commit()
@@ -161,20 +181,23 @@ async def place_order(user: Annotated[User, Depends(get_current_user)], session:
 @app.patch("/profile")
 async def update_profile(profile: ProfileUpdate, user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     try:
-        if re.match(r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$', profile.email) is None:
-            raise HTTPException(status_code=422, detail={"message": "Validation error", "email": "Validation error", })
+        if profile.avatar is None and profile.email is None and profile.password is None and profile.fio is None:
+            return JSONResponse(status_code=400, content={"message": "No input provided"})
 
-        if profile.password is None:
-            raise HTTPException(status_code=422, detail={"message": "Validation error", "password": "Validation error", })
+        if not validate_email(profile.email):
+            return generate_validation_error_for_fields("email")
 
-        if profile.email:
+        if profile.password == "":
+            return generate_validation_error_for_fields("password")
+
+        if profile.email is not None:
             await session.execute(text("update users set email = :email where id = :id"), {"email": profile.email, "id": user.id})
 
-        if profile.fio:
+        if profile.fio is not None:
             fio = profile.fio.split()
 
             if len(fio) < 2 or len(fio) > 3:
-                raise HTTPException(status_code=422, detail={"message": "Validation error", "fio": "Validation error", })
+                return generate_validation_error_for_fields("fio")
 
             name = fio[0]
             surname = fio[1]
@@ -200,8 +223,8 @@ async def update_profile(profile: ProfileUpdate, user: Annotated[User, Depends(g
 @app.get("/order")
 async def get_order_history(user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     try:
-        order_history = (await session.execute(text("select * from orders where user_id = :user_id"), {"user_id": user.id})).all()
-        return order_history
+        order_history = (await session.execute(text("select distinct id, (select array_agg(product_id) from order_items where order_id = orders.id), order_price as products from orders where user_id = :user_id"), {"user_id": user.id})).all()
+        return  [product._asdict() for product in order_history]
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
@@ -211,13 +234,12 @@ async def get_order_history(user: Annotated[User, Depends(get_current_user)], se
 @app.post("/product")
 async def create_product(product: Product, user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     if not user.is_admin:
-        raise HTTPException(status_code=403, detail="Forbidden")
+        return forbidden
 
     if math.isinf(product.price) or math.isnan(product.price):
-        raise HTTPException(status_code=400, detail={"message": "Validation error", "price": "Validation error"})
+        return generate_validation_error_for_fields("price")
     if product.name == "":
-        raise HTTPException(status_code=422, detail={"message": "Validation error", "password": "Validation error", })
-
+        return generate_validation_error_for_fields("name")
 
     try:
         product_id = (await session.execute(text("insert into products (name, price, user_id) values (:name, :price, :user_id) returning id"),
@@ -232,7 +254,7 @@ async def create_product(product: Product, user: Annotated[User, Depends(get_cur
 @app.delete("/product/{product_id}")
 async def delete_product(product_id: int, user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     if not user.is_admin:
-        raise HTTPException(status_code=403, detail="Forbidden")
+        return forbidden
 
     try:
         await session.execute(text("delete from products where id = :id"), {"id": product_id})
@@ -246,13 +268,13 @@ async def delete_product(product_id: int, user: Annotated[User, Depends(get_curr
 @app.patch("/product/{product_id}")
 async def update_product(product_id: int, product: ProductUpdate, user: Annotated[User, Depends(get_current_user)], session: SessionDep):
     if not user.is_admin:
-        raise HTTPException(status_code=403, detail="Forbidden")
+        return forbidden
 
-    if product.name is None and product.name and None or product.price and None:
-        raise HTTPException(status_code=422, detail={"message": "Validation error", "name": "Validation error", })
+    if product.name is None and product.name is None or product.price is None:
+        return JSONResponse(status_code=400, content={"message": "No input provided"})
 
     if product.price is not None and (math.isinf(product.price) or math.isnan(product.price)):
-        raise HTTPException(status_code=422, detail={"message": "Validation error", "price": "Validation error", })
+        return generate_validation_error_for_fields("price")
 
     try:
         if product.name is not None:
@@ -271,6 +293,6 @@ async def update_product(product_id: int, product: ProductUpdate, user: Annotate
     product_data = product_data.first()
 
     if product_data is None:
-        raise HTTPException(status_code=404, detail="Product not found")
+        return JSONResponse(status_code=404, content={"message": "Product not found"})
 
     return Product(**product_data._asdict())
